refactor: save GameInput’s created asset in the correct file
