+++ CommunityProject/Assets/_Game/Scripts/Runtime/Systems/InputSystem/ScriptableObjects/CallbackProcessors/GameplayActionsSO.cs
@@ -1,5 +1,4 @@
 using System;
-using BoundfoxStudios.CommunityProject.Input;
 using BoundfoxStudios.CommunityProject.Systems.SettingsSystem.ScriptableObjects;
 using UnityEngine;
 using UnityEngine.InputSystem;
