+++ CommunityProject/Assets/_Game/Settings/Input/GameInput.inputactions.meta
@@ -9,6 +9,6 @@ ScriptedImporter:
   assetBundleVariant: 
   script: {fileID: 11500000, guid: 8404be70184654265930450def6a9037, type: 3}
   generateWrapperCode: 1
-  wrapperCodePath: Assets/_Game/Scripts/Runtime/Input/GameInput.cs
+  wrapperCodePath: Assets/_Game/Scripts/Runtime/Systems/InputSystem/GameInput.cs
   wrapperClassName: 
-  wrapperCodeNamespace: BoundfoxStudios.CommunityProject.Input
+  wrapperCodeNamespace: BoundfoxStudios.CommunityProject.Systems.InputSystem
